+++ src/main/java/cn/com/sinofaith/bean/zfbBean/ZfbZhmxJylxEntity.java
@@ -1,6 +1,7 @@
 package cn.com.sinofaith.bean.zfbBean;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "zfbzhmx_jylx")
@@ -11,9 +12,9 @@ public class ZfbZhmxJylxEntity {
     private String xfmc;
     private long jyzcs;
     private long czzcs;
-    private double czzje;
+    private BigDecimal czzje;
     private long jzzcs;
-    private double jzzje;
+    private BigDecimal jzzje;
     private String insert_time;
     private long aj_id;
 
@@ -79,11 +80,11 @@ public class ZfbZhmxJylxEntity {
 
     @Basic
     @Column(name = "czzje",nullable = false,precision = 0)
-    public double getCzzje() {
+    public BigDecimal getCzzje() {
         return czzje;
     }
 
-    public void setCzzje(double czzje) {
+    public void setCzzje(BigDecimal czzje) {
         this.czzje = czzje;
     }
 
@@ -99,11 +100,11 @@ public class ZfbZhmxJylxEntity {
 
     @Basic
     @Column(name = "jzzje",nullable = false,precision = 0)
-    public double getJzzje() {
+    public BigDecimal getJzzje() {
         return jzzje;
     }
 
-    public void setJzzje(double jzzje) {
+    public void setJzzje(BigDecimal jzzje) {
         this.jzzje = jzzje;
     }
